Fixed all the image arithmetic functions. Added conjugate function and complex multiplication. Added arithmetic Tests. Fixed FFT code and added correlation test (not yet complete).

+++ src/kiss_fft.c
@@ -220,7 +220,7 @@ static void kf_bfly_generic(
     int u,k,q1,q;
     kiss_fft_cpx * twiddles = st->twiddles;
     kiss_fft_cpx t;
-    int Norig = st->nfft;
+    size_t Norig = st->nfft;
 
     CHECKBUF(scratchbuf,nscratchbuf,p);
 
@@ -234,7 +234,7 @@ static void kf_bfly_generic(
 
         k=u;
         for ( q1=0 ; q1<p ; ++q1 ) {
-            int twidx=0;
+            size_t twidx=0;
             Fout[ k ] = scratchbuf[0];
             for (q=1;q<p;++q ) {
                 twidx += fstride * k;
